Update README.md
This is a CT7201 class assessment that focuses on data manipulation, exploration, analysis, machine learning models and prediction on the Titanic dataset from Kaggle using python as the primary programming language completed in a Jupyter notebook. The result is using machine learning to create a model that predicts which group of passengers survived the infamous Titanic shipwreck.